refactor(projection): hoist pure county-aggregation math out of persistence
The WO-45 kind-dispatch surfaced a latent layering break: tui →
projection.{state,national,social_class} → persistence.county_aggregation
violated 'tui client reads projections only'. The three helpers involved
(_ideology_to_ternary, aggregate_survival_for_county,
aggregate_consciousness_for_county) are pure WorldState math and never
touch a database — relocated verbatim to babylon.projection.aggregation;
persistence re-exports unchanged so every bridge caller keeps its path.
NationalValueAggregate import in national.py was annotation-only → moved
under TYPE_CHECKING. import-linter: 9/9 contracts kept.
