+++ src/babylon/projection/county.py
@@ -56,7 +56,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from babylon.models.enums.topology import EdgeType, NodeType
-from babylon.persistence.county_aggregation import (
+from babylon.projection.aggregation import (
     aggregate_consciousness_for_county,
     aggregate_survival_for_county,
 )
